Rename scripture.py
