+++ verse.py
@@ -1,7 +1,7 @@
 """Select a random verse."""
 # pylint: disable=C0103
-import scripture
+import random_verse
 
-rand_verse = scripture.get_random_verse()
+rand_verse = random_verse.get_random_verse()
 print("Random verse: ")
-print(scripture.pretty_print_verse(rand_verse))
+print(random_verse.pretty_print_verse(rand_verse))
